Add a daybed serializer.
Also, this fixes a bit the serialisation process, as it wasn't serializing the
required attributes nor was able to do the relation between lists and the
different types of list displays we have in the form builder.

For instance radio buttons and dropdown have the same underlying structure, so
we were in need of a way to distinguish between them.

They now have a "formbuilderType" field in addition to the other ones, which
allows form builder to know which element to use for display.