Add a way to list the forms.
Here, we introduce a router to our react application.
We discovered along the way that the react-bootstrap weren't directly
compatible with react-router and had to use a specifc implementation of them
(and lost a couple hours in the meantime).

As the session was shared with a token in the hash part of the url, it was
coliding with how routing is done by react-router, so we're storing the daybed
session in localStorage for now (but this will probably evolve in the future).