Fix the deployment script.
Using uglify, for whatever reason doesn't work. We tried to hunt down the
problem but without success: it seems that uglify isn't able to parse some
file, but we were not able to tell which one.

Because we still need to deploy new versions of daybed, I've deactivated the
minification process for now, until we figure out. See
https://github.com/spiral-project/formbuilder/issues/27