Make the menu affix.
In order to have something that works well when doing the affix, I needed to
"hardcode" the fluid grid sized into the CSS.

It seems to be the way to go, even if I find it a bit hard to maintain. Maybe
using a CSS pre-processor would make things easier to handle here.